refactor: extract shared render building blocks into render-shared.mjs (tc-nnte.7)
